Fix notifications (role-specific) + real-time new tasks
- Layout.jsx: replace useLocalData with useData (Supabase)
- Notifications now filtered per user role:
  Admin sees all alerts, each worker sees only their own
- Fix INTEGER overflow: change id columns to BIGINT so
  Date.now() IDs from new tasks sync correctly to Supabase

+++ aquaops/supabase_setup.sql
@@ -15,27 +15,27 @@ DROP TABLE IF EXISTS users;
 -- ============================================================
 
 CREATE TABLE users (
-  id INTEGER PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   data JSONB NOT NULL
 );
 
 CREATE TABLE customers (
-  id INTEGER PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   data JSONB NOT NULL
 );
 
 CREATE TABLE meter_readings (
-  id INTEGER PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   data JSONB NOT NULL
 );
 
 CREATE TABLE tickets (
-  id INTEGER PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   data JSONB NOT NULL
 );
 
 CREATE TABLE tasks (
-  id INTEGER PRIMARY KEY,
+  id BIGINT PRIMARY KEY,
   data JSONB NOT NULL
 );
